add balance tracking
