+++ x/gov/keeper/common_test.go
@@ -3,21 +3,24 @@ package keeper_test
 import (
 	"testing"
 
-	"github.com/stretchr/testify/require"
+	"cosmossdk.io/math"
+	"github.com/golang/mock/gomock"
+	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
+	tmtime "github.com/tendermint/tendermint/types/time"
 
-	simtestutil "github.com/cosmos/cosmos-sdk/testutil/sims"
+	"github.com/cosmos/cosmos-sdk/baseapp"
+	"github.com/cosmos/cosmos-sdk/testutil"
 	"github.com/cosmos/cosmos-sdk/testutil/testdata"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	moduletestutil "github.com/cosmos/cosmos-sdk/types/module/testutil"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
+	"github.com/cosmos/cosmos-sdk/x/gov/keeper"
 	govtestutil "github.com/cosmos/cosmos-sdk/x/gov/testutil"
 	"github.com/cosmos/cosmos-sdk/x/gov/types"
 	v1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1"
 	"github.com/cosmos/cosmos-sdk/x/gov/types/v1beta1"
-	"github.com/cosmos/cosmos-sdk/x/staking"
-	stakingkeeper "github.com/cosmos/cosmos-sdk/x/staking/keeper"
-	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
+	minttypes "github.com/cosmos/cosmos-sdk/x/mint/types"
 )
 
 var (
@@ -38,50 +41,71 @@ func getTestProposal() []sdk.Msg {
 	}
 }
 
-func createValidators(
-	t *testing.T,
-	acctKeeper govtestutil.AccountKeeper,
-	bankKeeper govtestutil.BankKeeper,
-	stakingKeeper govtestutil.StakingKeeper,
-	ctx sdk.Context,
-	powers []int64,
-) ([]sdk.AccAddress, []sdk.ValAddress) {
-	addrs := simtestutil.AddTestAddrsIncremental(bankKeeper, stakingKeeper, ctx, 5, sdk.NewInt(30000000))
-	valAddrs := simtestutil.ConvertAddrsToValAddrs(addrs)
-	pks := simtestutil.CreateTestPubKeys(5)
-	cdc := moduletestutil.MakeTestEncodingConfig().Codec
+func setupGovKeeper(t *testing.T) (
+	*keeper.Keeper,
+	*govtestutil.MockAccountKeeper,
+	*govtestutil.MockBankKeeper,
+	*govtestutil.MockStakingKeeper,
+	moduletestutil.TestEncodingConfig,
+	sdk.Context,
+) {
+	key := sdk.NewKVStoreKey(types.StoreKey)
+	testCtx := testutil.DefaultContextWithDB(t, key, sdk.NewTransientStoreKey("transient_test"))
+	ctx := testCtx.Ctx.WithBlockHeader(tmproto.Header{Time: tmtime.Now()})
+	encCfg := moduletestutil.MakeTestEncodingConfig()
+	v1.RegisterInterfaces(encCfg.InterfaceRegistry)
+	v1beta1.RegisterInterfaces(encCfg.InterfaceRegistry)
+	banktypes.RegisterInterfaces(encCfg.InterfaceRegistry)
 
-	// Create a new real (not mocked) staking keeper just in this function.
-	sk := stakingkeeper.NewKeeper(
-		cdc,
-		sdk.NewKVStoreKey(stakingtypes.StoreKey),
-		acctKeeper,
-		bankKeeper,
-		authtypes.NewModuleAddress(types.ModuleName).String(),
-	)
+	// Create MsgServiceRouter, but don't populate it before creating the gov
+	// keeper.
+	msr := baseapp.NewMsgServiceRouter()
 
-	val1, err := stakingtypes.NewValidator(valAddrs[0], pks[0], stakingtypes.Description{})
-	require.NoError(t, err)
-	val2, err := stakingtypes.NewValidator(valAddrs[1], pks[1], stakingtypes.Description{})
-	require.NoError(t, err)
-	val3, err := stakingtypes.NewValidator(valAddrs[2], pks[2], stakingtypes.Description{})
-	require.NoError(t, err)
+	// gomock initializations
+	ctrl := gomock.NewController(t)
+	acctKeeper := govtestutil.NewMockAccountKeeper(ctrl)
+	bankKeeper := govtestutil.NewMockBankKeeper(ctrl)
+	stakingKeeper := govtestutil.NewMockStakingKeeper(ctrl)
+	acctKeeper.EXPECT().GetModuleAddress(types.ModuleName).Return(govAcct).AnyTimes()
+	acctKeeper.EXPECT().GetModuleAccount(gomock.Any(), types.ModuleName).Return(authtypes.NewEmptyModuleAccount(types.ModuleName)).AnyTimes()
+	trackMockBalances(bankKeeper)
+	stakingKeeper.EXPECT().TokensFromConsensusPower(ctx, gomock.Any()).DoAndReturn(func(ctx sdk.Context, power int64) math.Int {
+		return sdk.TokensFromConsensusPower(power, math.NewIntFromUint64(1000000))
+	}).AnyTimes()
+	stakingKeeper.EXPECT().BondDenom(ctx).Return("stake").AnyTimes()
 
-	sk.SetValidator(ctx, val1)
-	sk.SetValidator(ctx, val2)
-	sk.SetValidator(ctx, val3)
-	sk.SetValidatorByConsAddr(ctx, val1)
-	sk.SetValidatorByConsAddr(ctx, val2)
-	sk.SetValidatorByConsAddr(ctx, val3)
-	sk.SetNewValidatorByPowerIndex(ctx, val1)
-	sk.SetNewValidatorByPowerIndex(ctx, val2)
-	sk.SetNewValidatorByPowerIndex(ctx, val3)
+	// Gov keeper initializations
+	govKeeper := keeper.NewKeeper(encCfg.Codec, key, acctKeeper, bankKeeper, stakingKeeper, msr, types.DefaultConfig(), govAcct.String())
+	govKeeper.SetProposalID(ctx, 1)
+	govRouter := v1beta1.NewRouter() // Also register legacy gov handlers to test them too.
+	govRouter.AddRoute(types.RouterKey, v1beta1.ProposalHandler)
+	govKeeper.SetLegacyRouter(govRouter)
+	govKeeper.SetParams(ctx, v1.DefaultParams())
 
-	_, _ = sk.Delegate(ctx, addrs[0], stakingKeeper.TokensFromConsensusPower(ctx, powers[0]), stakingtypes.Unbonded, val1, true)
-	_, _ = sk.Delegate(ctx, addrs[1], stakingKeeper.TokensFromConsensusPower(ctx, powers[1]), stakingtypes.Unbonded, val2, true)
-	_, _ = sk.Delegate(ctx, addrs[2], stakingKeeper.TokensFromConsensusPower(ctx, powers[2]), stakingtypes.Unbonded, val3, true)
+	// Register all handlers for the MegServiceRouter.
+	msr.SetInterfaceRegistry(encCfg.InterfaceRegistry)
+	v1.RegisterMsgServer(msr, keeper.NewMsgServerImpl(govKeeper))
+	banktypes.RegisterMsgServer(msr, nil) // Nil is fine here as long as we never execute the proposal's Msgs.
 
-	_ = staking.EndBlocker(ctx, sk)
+	return govKeeper, acctKeeper, bankKeeper, stakingKeeper, encCfg, ctx
+}
 
-	return addrs, valAddrs
+func trackMockBalances(bankKeeper *govtestutil.MockBankKeeper) {
+	balances := make(map[string]sdk.Coins)
+	bankKeeper.EXPECT().MintCoins(gomock.Any(), minttypes.ModuleName, gomock.Any()).AnyTimes()
+	bankKeeper.EXPECT().BurnCoins(gomock.Any(), types.ModuleName, gomock.Any()).AnyTimes()
+	bankKeeper.EXPECT().SendCoinsFromModuleToModule(gomock.Any(), minttypes.ModuleName, types.ModuleName, gomock.Any()).AnyTimes()
+
+	bankKeeper.EXPECT().SendCoinsFromAccountToModule(gomock.Any(), gomock.Any(), types.ModuleName, gomock.Any()).DoAndReturn(func(_ sdk.Context, sender sdk.AccAddress, _ string, coins sdk.Coins) error {
+		balances[sender.String()] = balances[sender.String()].Sub(coins...)
+		return nil
+	}).AnyTimes()
+	bankKeeper.EXPECT().SendCoinsFromModuleToAccount(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).DoAndReturn(func(_ sdk.Context, module string, rcpt sdk.AccAddress, coins sdk.Coins) error {
+		balances[rcpt.String()] = balances[rcpt.String()].Add(coins...)
+		return nil
+	}).AnyTimes()
+	bankKeeper.EXPECT().GetAllBalances(gomock.Any(), gomock.Any()).DoAndReturn(func(_ sdk.Context, addr sdk.AccAddress) sdk.Coins {
+		return balances[addr.String()]
+	}).AnyTimes()
+
 }
